+++ shortcode/search-properties-map.php
@@ -34,7 +34,7 @@ function rentfetch_propertymap( $atts ) {
     
     // the map itself
     $key = get_field( 'google_maps_api_key', 'option' );
-    wp_enqueue_script( 'rentfetch-google-maps', 'https://maps.googleapis.com/maps/api/js?key=' . $key, array(), RENTFETCH_VERSION, true );
+    wp_enqueue_script( 'rentfetch-google-maps', 'https://maps.googleapis.com/maps/api/js?key=' . $key, array(), null, true );
     
     // Localize the google maps script, then enqueue that
     $maps_options = array(