+++ template/single-properties-functions.php
@@ -338,7 +338,7 @@ function rentfetch_single_property_map() {
 
     // the map itself
     $key = get_field( 'google_maps_api_key', 'option' );
-    wp_enqueue_script( 'rentfetch-google-maps', 'https://maps.googleapis.com/maps/api/js?key=' . $key, array(), RENTFETCH_VERSION, true );
+    wp_enqueue_script( 'rentfetch-google-maps', 'https://maps.googleapis.com/maps/api/js?key=' . $key, array(), null, true );
 
     // Localize the google maps script, then enqueue that
     $maps_options = array(